Rearranged original rockets-starter into files
