+++ packages/rockets-web/src/app/screens/Home/Home.tsx
@@ -1,0 +1,21 @@
+import React from "react";
+
+const Home = () => {
+  return (
+    <div
+      style={{
+        height: "100%",
+        width: "100%",
+        display: "flex",
+        flexDirection: "column",
+        justifyContent: "center",
+        alignItems: "center",
+        color: "white",
+      }}
+    >
+      LOGGED IN!
+    </div>
+  );
+};
+
+export default Home;
