+++ packages/rockets-web/src/app/screens/NotFound/NotFound.tsx
@@ -1,0 +1,21 @@
+import React from "react";
+
+const NotFound = () => {
+  return (
+    <div
+    // style={{
+    //   height: "100%",
+    //   width: "100%",
+    //   display: "flex",
+    //   flexDirection: "column",
+    //   justifyContent: "center",
+    //   alignItems: "center",
+    //   color: "white",
+    // }}
+    >
+      Not Found
+    </div>
+  );
+};
+
+export default NotFound;
