+++ packages/rockets-web/src/app/screens/Unauthorized/Unauthorized.tsx
@@ -1,0 +1,21 @@
+import React from "react";
+
+const Unauthorized = () => {
+  return (
+    <div
+    // style={{
+    //   height: "100%",
+    //   width: "100%",
+    //   display: "flex",
+    //   flexDirection: "column",
+    //   justifyContent: "center",
+    //   alignItems: "center",
+    //   color: "white",
+    // }}
+    >
+      Unauthorized
+    </div>
+  );
+};
+
+export default Unauthorized;
