+++ packages/rockets-web/src/app/screens/Unauthorized/index.ts
@@ -1,0 +1,3 @@
+import Unauthorized from "./Unauthorized";
+
+export default Unauthorized;
